+++ packages/alpha/src/types.ts
@@ -1,270 +1,142 @@
 // Copyright 2020 Cognite AS
 
-import {
-  Metadata,
-  MetadataPatch,
-  SinglePatchRequiredString,
-  SinglePatchString,
-  Timestamp,
-} from '@cognite/sdk';
+import { SinglePatch, Timestamp } from '@cognite/sdk';
 import {
   CogniteExternalId,
   CogniteInternalId,
-  ExternalId,
   FilterQuery,
   InternalId,
 } from '@cognite/sdk-core';
 
 export * from '@cognite/sdk';
 
-export type MonitoringTaskModelExternalId = 'threshold' | 'double_threshold';
+// This file is here mostly to allow apis to import { ... } from '../../types';
+// Overriding types should probably be done in their respective API endpoint files, where possible
 
-export const MonitoringTaskModelExternalId = {
-  THRESHOLD: 'threshold' as const,
-  DOUBLE_THRESHOLD: 'double_threshold' as const,
-};
-
-export interface MonitoringTaskThresholdModelCreateBase {
-  externalId: MonitoringTaskModelExternalId;
+export interface SimulatorUnitRecord {
+  label: string;
+  value: string;
 }
-export interface MonitoringTaskThresholdModelCreate
-  extends MonitoringTaskThresholdModelCreateBase {
-  externalId: 'threshold';
-  timeseriesExternalId?: CogniteExternalId;
-  timeseriesId?: CogniteInternalId;
-  threshold: number;
-  granularity?: string;
+export interface SimulatorUnitSystem {
+  label: string;
+  defaultUnits: Record<string, string>;
 }
 
-export interface MonitoringTaskDoubleThresholdModelCreate
-  extends MonitoringTaskThresholdModelCreateBase {
-  externalId: 'double_threshold';
-  timeseriesExternalId?: CogniteExternalId;
-  timeseriesId?: CogniteInternalId;
-  lowerThreshold?: number;
-  upperThreshold?: number;
-  granularity?: string;
+export interface SimulatorUnitsMap {
+  label: string;
+  units: SimulatorUnitRecord[];
 }
 
-export interface MonitoringTaskThresholdModel {
-  externalId: 'threshold';
-  timeseriesId: CogniteInternalId;
-  granularity?: string;
-  threshold: number;
+export interface SimulatorUnits {
+  unitsMap?: Record<string, SimulatorUnitsMap>;
+  unitSystem?: Record<string, SimulatorUnitSystem>;
 }
 
-export interface MonitoringTaskDoubleThresholdModel {
-  externalId: 'double_threshold';
-  timeseriesId: CogniteInternalId;
-  timeseriesExternalId?: CogniteExternalId;
-  granularity?: string;
-  lowerThreshold?: number;
-  upperThreshold?: number;
+export interface SimulatorBoundaryCondition {
+  name: string;
+  address: string;
+  key: string;
 }
 
-export interface MonitoringTaskCreate {
+export interface SimulatorModelType {
+  name: string;
+  key: string;
+}
+
+export interface SimulatorStepField {
+  name: string;
+  label: string;
+  info: string;
+}
+
+export interface SimulatorStep {
+  stepType: string;
+  fields: SimulatorStepField[];
+}
+export interface Simulator {
+  id: CogniteInternalId;
   externalId: CogniteExternalId;
-  name: string;
-  channelId: CogniteInternalId;
-  interval: number;
-  overlap: number;
-  model:
-    | MonitoringTaskThresholdModelCreate
-    | MonitoringTaskDoubleThresholdModelCreate;
-  nonce: string;
-  source?: string;
-  sourceId?: string;
+  fileExtensionTypes: string[];
+  isBoundaryConditionsEnabled: boolean;
+  boundaryConditions: SimulatorBoundaryCondition[];
+  isCalculationsEnabled: boolean;
+  modelTypes: SimulatorModelType[];
+  enabled: boolean;
+  stepFields?: SimulatorStep[];
+  units?: SimulatorUnits;
+  createdTime: Timestamp;
+  lastUpdatedTime: Timestamp;
 }
 
-export interface MonitoringTask {
+export interface SimulatorFilter {
+  enabled?: boolean;
+}
+
+export interface SimulatorFilterQuery extends FilterQuery {
+  filter?: SimulatorFilter;
+}
+
+export interface SimulatorCreate {
+  externalId: CogniteExternalId;
+  fileExtensionTypes: string[];
+  isBoundaryConditionsEnabled?: boolean;
+  boundaryConditions?: SimulatorBoundaryCondition[];
+  isCalculationsEnabled?: boolean;
+  modelTypes?: SimulatorModelType[];
+  enabled?: boolean;
+  stepFields?: SimulatorStep[];
+  units?: SimulatorUnits;
+}
+
+export interface SimulatorIntegration {
   id: CogniteInternalId;
   externalId?: CogniteExternalId;
-  name: string;
-  channelId: CogniteInternalId;
-  interval: number;
-  overlap: number;
-  model: MonitoringTaskThresholdModel | MonitoringTaskDoubleThresholdModel;
-  source?: string;
-  sourceId?: string;
+  simulatorExternalId?: CogniteExternalId;
+  heartbeat: Timestamp;
+  dataSetId: CogniteInternalId;
+  connectorVersion: string;
+  simulatorVersion: string;
+  runApiEnabled: boolean;
+  licenseStatus: string;
+  licenseLastCheckedTime: Timestamp;
+  connectorStatus: string;
+  connectorStatusUpdatedTime: Timestamp;
+  createdTime: Timestamp;
+  lastUpdatedTime: Timestamp;
 }
 
-export interface MonitoringTaskModelFilter {
-  timeseriesId: CogniteInternalId;
-}
-
-export interface MonitoringTaskFilter {
-  externalIds?: CogniteExternalId[];
-  ids?: CogniteInternalId[];
-  channelIds?: CogniteInternalId[];
-  model?: MonitoringTaskModelFilter;
-}
-
-export interface MonitoringTaskFilterQuery extends FilterQuery {
-  filter?: MonitoringTaskFilter;
-}
-
-export interface AlertFilter {
-  channelIds?: CogniteInternalId[];
-  channelExternalIds?: CogniteExternalId[];
-}
-
-export interface AlertFilterQuery extends FilterQuery {
-  filter?: AlertFilter;
-}
-
-export interface AlertDeduplicationRuleCreate {
-  mergeInterval?: string;
-  activationInterval?: string;
-}
-
-export interface AlertRulesCreate {
-  deduplication: AlertDeduplicationRuleCreate;
-}
-
-export interface AlertDeduplicationRule {
-  mergeInterval?: string;
-  activationInterval?: string;
-}
-
-export interface AlertRules {
-  deduplication: AlertDeduplicationRule;
-}
-
-export interface AlertCreate {
+export interface SimulatorIntegrationCreate {
   externalId?: CogniteExternalId;
-  timestamp?: Timestamp;
-  channelId?: CogniteInternalId;
-  channelExternalId?: CogniteExternalId;
-  source: string;
-  value?: string;
-  level?: 'DEBUG' | 'INFO' | 'WARNING' | 'ERROR' | 'CRITICAL';
-  metadata?: Metadata;
+  simulatorExternalId?: CogniteExternalId;
+  dataSetId?: CogniteInternalId;
+  connectorVersion?: string;
+  simulatorVersion?: string;
+  runApiEnabled?: boolean;
+  licenseStatus?: string;
+  licenseLastCheckedTime?: Timestamp;
+  connectorStatus?: string;
+  connectorStatusUpdatedTime?: Timestamp;
 }
 
-export interface Alert {
-  id: CogniteInternalId;
-  externalId?: CogniteExternalId;
-  timestamp: Timestamp;
-  channelId: CogniteInternalId;
-  channelExternalId?: CogniteExternalId;
-  source: string;
-  value?: string;
-  level?: 'DEBUG' | 'INFO' | 'WARNING' | 'ERROR' | 'CRITICAL';
-  metadata?: Metadata;
-  acknowledged: boolean;
-  closed: boolean;
+export interface SimulatorIntegrationFilter {
+  simulatorExternalIds?: CogniteExternalId[];
 }
 
-export interface ChannelCreate {
-  externalId?: CogniteExternalId;
-  parentExternalId?: CogniteExternalId;
-  parentId?: CogniteInternalId;
-  name: string;
-  description?: string;
-  metadata?: Metadata;
-  alertRules?: AlertRulesCreate;
+export interface SimulatorIntegrationFilterQuery extends FilterQuery {
+  filter?: SimulatorIntegrationFilter;
 }
 
-export interface Channel {
-  id: CogniteInternalId;
-  externalId?: CogniteExternalId;
-  parentExternalId?: CogniteExternalId;
-  parentId?: CogniteInternalId;
-  name: string;
-  description?: string;
-  metadata?: Metadata;
-  alertRules?: AlertRules;
-}
-
-export interface ChannelFilter {
-  ids?: CogniteInternalId[];
-  externalIds?: CogniteExternalId[];
-  parentIds?: CogniteInternalId[];
-  metadata?: Metadata;
-}
-
-export interface ChannelFilterQuery extends FilterQuery {
-  filter?: ChannelFilter;
-}
-
-export interface ChannelPatch {
+export interface SimulatorPatch {
   update: {
-    externalId?: SinglePatchString;
-    name?: SinglePatchRequiredString;
-    description?: SinglePatchString;
-    metadata?: MetadataPatch;
+    fileExtensionTypes?: SinglePatch<string[]>;
+    isBoundaryConditionsEnabled?: SinglePatch<boolean>;
+    boundaryConditions?: SinglePatch<SimulatorBoundaryCondition[]>;
+    isCalculationsEnabled?: SinglePatch<boolean>;
+    modelTypes?: SinglePatch<SimulatorModelType[]>;
+    enabled?: SinglePatch<boolean>;
+    stepFields?: SinglePatch<SimulatorStep[]>;
+    units?: SinglePatch<SimulatorUnits>;
   };
 }
 
-export interface ChannelChangeById extends ChannelPatch, InternalId {}
-export interface ChannelChangeByExternalId extends ChannelPatch, ExternalId {}
-export type ChannelChange = ChannelChangeById | ChannelChangeByExternalId;
-
-export interface SubscriberCreate {
-  externalId?: CogniteExternalId;
-  metadata?: Metadata;
-  email: string;
-}
-
-export interface SubscriberFilter {
-  externalIds?: CogniteExternalId[];
-  ids?: CogniteInternalId[];
-  metadata?: Metadata;
-  email?: string;
-}
-
-export interface SubscriberFilterQuery extends FilterQuery {
-  filter?: SubscriberFilter;
-}
-
-export interface Subscriber {
-  id: CogniteInternalId;
-  externalId?: CogniteExternalId;
-  metadata?: Metadata;
-  email: string;
-}
-
-export interface SubscriberCreate {
-  externalId?: CogniteExternalId;
-  metadata?: Metadata;
-  email: string;
-}
-
-export interface SubscriptionCreate {
-  externalId?: CogniteExternalId;
-  channelId?: CogniteInternalId;
-  channelExternalId?: CogniteExternalId;
-  subscriberId?: CogniteInternalId;
-  subscriberExternalId?: CogniteExternalId;
-  metadata?: Metadata;
-}
-
-export interface SubscriptionFilter {
-  externalIds?: CogniteExternalId[];
-  channelIds?: CogniteInternalId[];
-  channelExternalIds?: CogniteExternalId[];
-  subscriberIds?: CogniteInternalId[];
-  subscriberExternalIds?: CogniteExternalId[];
-  metadata?: Metadata;
-}
-
-export interface SubscriptionFilterQuery extends FilterQuery {
-  filter?: SubscriptionFilter;
-}
-
-export interface SubscriptionDelete {
-  externalId?: CogniteExternalId;
-  channelExternalId?: CogniteExternalId;
-  channelId?: CogniteInternalId;
-  subscriberId?: CogniteInternalId;
-  subscriberExternalId?: CogniteExternalId;
-}
-
-export interface Subscription {
-  externalId?: CogniteExternalId;
-  channelId: CogniteInternalId;
-  channelExternalId?: CogniteExternalId;
-  subscriberId: CogniteInternalId;
-  subscriberExternalId?: CogniteExternalId;
-  metadata?: Metadata;
-}
+export interface SimulatorChange extends SimulatorPatch, InternalId {}